+++ apps/tergite-dashboard/src/app.tsx
@@ -1,23 +1,36 @@
-import { useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { AppState } from "../types";
 import { newAppState, AppStateContext } from "./lib/app-state";
 import { AppRouter } from "./router";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { createBrowserRouter } from "react-router-dom";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
+import { getSavedIsDarkMode, saveIsDarkMode } from "./lib/utils";
 
 const queryClient = new QueryClient();
 
 export function App({ routerConstructor = createBrowserRouter }: Props) {
   const [currentProject, setCurrentProject] = useState<string>();
   const [apiToken, setApiToken] = useState<string>();
+  const [isDark, _setIsDark] = useState<boolean>(getSavedIsDarkMode());
+  const toggleIsDark = useCallback(
+    () => _setIsDark(!isDark),
+    [isDark, _setIsDark]
+  );
+
   const state: AppState = {
     ...newAppState(),
     apiToken,
     setApiToken,
+    isDark,
+    toggleIsDark,
     currentProjectExtId: currentProject,
     setCurrentProjectExtId: setCurrentProject,
   };
+
+  useEffect(() => {
+    saveIsDarkMode(isDark);
+  }, [isDark]);
 
   return (
     <QueryClientProvider client={queryClient}>